Change plugin package
