+++ intellij-plugin/src/main/java/org/quiltmc/intellij/mapping/language/QuiltMappingSyntaxHighlighterFactory.java
@@ -1,4 +1,4 @@
-package org.quiltmc.mapping.intellij.language;
+package org.quiltmc.intellij.mapping.language;
 
 import com.intellij.openapi.fileTypes.SyntaxHighlighter;
 import com.intellij.openapi.fileTypes.SyntaxHighlighterFactory;
