+++ intellij-plugin/src/main/resources/META-INF/plugin.xml
@@ -26,20 +26,20 @@
 	<extensions defaultExtensionNs="com.intellij">
 		<fileType
 				name="Quilt Mapping"
-				implementationClass="org.quiltmc.mapping.intellij.language.QuiltMappingFileType"
+				implementationClass="org.quiltmc.intellij.mapping.language.QuiltMappingFileType"
 				fieldName="INSTANCE"
 				language="QuiltMapping"
 				extensions="quiltmapping"/>
 		<lang.foldingBuilder
 			language="QuiltMapping"
-			implementationClass="org.quiltmc.mapping.intellij.language.QuiltMappingFoldingBuilder"/>
+			implementationClass="org.quiltmc.intellij.mapping.language.QuiltMappingFoldingBuilder"/>
 		<lang.parserDefinition
 			language="QuiltMapping"
-			implementationClass="org.quiltmc.mapping.intellij.language.QuiltMappingParserDefinition"/>
+			implementationClass="org.quiltmc.intellij.mapping.language.QuiltMappingParserDefinition"/>
 		<lang.syntaxHighlighterFactory
 			language="QuiltMapping"
-			implementationClass="org.quiltmc.mapping.intellij.language.QuiltMappingSyntaxHighlighterFactory"/>
+			implementationClass="org.quiltmc.intellij.mapping.language.QuiltMappingSyntaxHighlighterFactory"/>
 		<colorSettingsPage
-				implementation="org.quiltmc.mapping.intellij.language.QuiltMappingColorSettingsPage"/>
+				implementation="org.quiltmc.intellij.mapping.language.QuiltMappingColorSettingsPage"/>
 	</extensions>
 </idea-plugin>
